+++ packages/wekan-oidc/oidc_server.js
@@ -1,3 +1,5 @@
+import {addGroups, addEmail,changeFullname, changeUsername} from './loginHandler';
+
 Oidc = {};
 httpCa = false;
 
@@ -16,6 +18,9 @@ if (process.env.OAUTH2_CA_CERT !== undefined) {
 OAuth.registerService('oidc', 2, null, function (query) {
 
   var debug = process.env.DEBUG || false;
+  console.log(process.env);
+  var propagateOidcData = process.env.PROPAGATE_OIDC_DATA || false;
+
   var token = getToken(query);
   if (debug) console.log('XXX: register token:', token);
 
@@ -73,6 +78,20 @@ OAuth.registerService('oidc', 2, null, function (query) {
   var profile = {};
   profile.name = userinfo[process.env.OAUTH2_FULLNAME_MAP]; // || userinfo["displayName"];
   profile.email = userinfo[process.env.OAUTH2_EMAIL_MAP]; // || userinfo["email"];
+  if (propagateOidcData)
+  {
+    users= Meteor.users;
+    user = users.findOne({'services.oidc.id':  serviceData.id});
+    if(user)
+    {
+      serviceData.groups = profile.groups
+      profile.groups = userinfo["groups"];
+      if(userinfo["groups"]) addGroups(user, userinfo["groups"]);
+      if(profile.email) addEmail(user, profile.email)
+      if(profile.name) changeFullname(user, profile.name)
+      if(profile.username) changeUsername(user, profile.username)
+    }
+  }
   if (debug) console.log('XXX: profile:', profile);
 
   return {